gallery-v2/backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
motor==3.3.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
aiofiles==23.2.1
python-dotenv==1.0.1
httpx==0.27.0
Pillow==10.2.0
pydantic-settings==2.1.0
qrcode==7.4.2
